requirements.txt
funcsigs==1.0.2
kafka-python==1.4.4
mock==2.0.0
pbr==5.1.1
py4j==0.10.6
pyspark==2.3.0
pytz==2018.7
six==1.11.0
